initializeSdk optimize + minSdkVersion 17 + custom flush interval

+++ business-core/src/main/java/com/tiktok/appevents/TTRequest.java
@@ -169,7 +169,7 @@ class TTRequest {
             } else {
                 try {
                     JSONObject resultJson = new JSONObject(result);
-                    int code = (Integer) resultJson.getInt("code");
+                    int code = resultJson.getInt("code");
 
                     if (code == TTConst.ApiErrorCodes.API_ERROR.code) {
                         failedEventsToBeDiscarded.addAll(currentBatch);
